corrección get de mesa

+++ Free Hold'em/src/Servidor/Mesa.java
@@ -46,11 +46,11 @@ public class Mesa {
         this.apuestaRonda = apuestaRonda;
     }
 
-    public long getApuestaprevia() {
+    public int getApuestaprevia() {
         return apuestaprevia;
     }
 
-    public long getApuestaRonda() {
+    public int getApuestaRonda() {
         return apuestaRonda;
     }
